requirements.txt
# 数字人视频生成器 - Python 依赖
# 安装命令: pip install -r requirements.txt

# 核心依赖
numpy>=1.21.0
opencv-python>=4.5.0
soundfile>=0.12.0

# PyTorch (VAD 人声检测)
# 如需 GPU 加速，请访问 https://pytorch.org 选择对应版本
torch>=2.0.0
torchaudio>=2.0.0

# Gemini API (可选，用于自动分析视频)
google-genai>=0.3.0

# 注意: FFmpeg 需要单独安装
# Windows: choco install ffmpeg 或 scoop install ffmpeg
# macOS: brew install ffmpeg
# Linux: sudo apt install ffmpeg
